% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/distance_neighbor.R
\name{calculate_neighbors}
\alias{calculate_neighbors}
\title{calculate_neighbors function}
\usage{
calculate_neighbors(distMatrix)
}
\arguments{
\item{distMatrix}{A distance matrix between data points.}
}
\value{
A matrix which includes, for each data point, column index to the
 ordered nearest neighbors. Note that first row is just de data point index
 into consideration.
}
\description{
For each point, sort the other points according to a distance matrix to find
a sortest list of points ordered by distance (nearest neighbors).
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(200), ncol=2)
inputsTrain  <- matrix(rnorm(200), ncol=2)
classesTrain <- c(rep(0, 50), rep(1, 50))
# compute distances
distMatrix<-distance(inputsTest, memory=inputsTrain,p=2,type="predict")
# compute neighbors
neighbors<-calculate_neighbors(distMatrix)
}

