% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/distance_neighbor.R
\name{distance}
\alias{distance}
\title{distance function}
\usage{
distance(data, memory = NULL, p = 2, type = "train")
}
\arguments{
\item{data}{A data frame or a matrix where rows are observations and
columns are features. If \code{type} is "train" this is training
dataset, and if it is "predict" it is test dataset.}

\item{memory}{A data frame or a matrix where rows are observations
and columns are features. If \code{type} is "train" this argument
is not needed, and if it is "predict" it is a training dataset.}

\item{p}{Distance metric the classifier should use, the value can be
either 1, 2 or Inf.}

\item{type}{Whether the goal is to train the classifier or predict
classes of new observations based on past ones. The value can be
either "train" or "predict".}
}
\value{
If \code{type} is "train", a squared symmetric matrix whose i,j value is the is the distance between
data points i and j. Note that i,j are data observations related to same rows in
\code{data} file. If \code{type} is "predict", a matrix with as many rows as
observations in \code{data} and as many columns as observations in
\code{memory}; value i,j is distance between data[i] and memory[j].
}
\description{
Calculate the distances between different input data.
}
\examples{
inputsTest   <- matrix(rnorm(200), ncol=2)
inputsTrain  <- matrix(rnorm(200), ncol=2)
# compute distances
distMatrix<-distance(inputsTest, memory=inputsTrain,p=2,type="predict")
}

